Rename files and comments
Use example instead of using Boilerplate.
