+++ account/update_account_with_account_key_public/index.js
@@ -14,9 +14,9 @@ let senderPrivateKey = '' // e.g. '0x39a6375b608c2572fadb2ed9fd78c5c456ca3aa860c
 let recipientAddress = '' // e.g. '0xeb709d59954f4cdc6b6f3bfcd8d531887b7bd199'
 
 /**
- * Boilerplate code about "Using IPFS with Caver."
- * Related article - Korean: https://medium.com/klaytn/caver%EB%A5%BC-%EC%9D%B4%EC%9A%A9%ED%95%9C-ipfs-%EC%82%AC%EC%9A%A9%EB%B2%95-4889a3b29c0b
- * Related article - English: https://medium.com/klaytn/using-ipfs-with-caver-964e1f721bfe
+ * Example code about "How to Update Klaytn Account Keys with Caver #1 — AccountKeyPublic"
+ * Related article - Korean: https://medium.com/klaytn/caver-caver%EB%A1%9C-klaytn-%EA%B3%84%EC%A0%95%EC%9D%98-%ED%82%A4%EB%A5%BC-%EB%B0%94%EA%BE%B8%EB%8A%94-%EB%B0%A9%EB%B2%95-1-accountkeypublic-7f8a7197e2d4
+ * Related article - English: https://medium.com/klaytn/caver-how-to-update-klaytn-account-keys-with-caver-1-accountkeypublic-30336b8f0b50
  */
 async function main() {
     try {
@@ -45,6 +45,7 @@ function loadEnv() {
 }
 
 async function run() {
+    console.log(`=====> Update AccountKey to AccountKeyPublic`)
     const option = {
         headers: [
             {
@@ -56,44 +57,55 @@ async function run() {
     }
     const caver = new Caver(new Caver.providers.HttpProvider(nodeApiUrl, option))
 
-    // Set connection with IPFS Node
-    caver.ipfs.setIPFSNode('ipfs.infura.io', 5001, true)
-    // `ipfs.txt` is located at `caver-js-examples/ipfs/using_ipfs_with_caver/resources`.
-    const testFile = `${__dirname}/resources/ipfs.txt`
-
-    // Add a file to IPFS with file path
-    const cid = await caver.ipfs.add(testFile)
-    console.log(`cid: ${cid}`)
-
-    // // Add a file to IPFS with file contents
-    // const contents = Buffer.from('test data')
-    // const cid = await caver.ipfs.add(contents)
-
-    // Download a file from IPFS
-    const buffer = await caver.ipfs.get(cid)
-    console.log(`Contents downloaded from IPFS: ${buffer.toString('utf8')}`)
-
-    // Convert from CID to multihash(hex formatted)
-    const multihash = await caver.ipfs.toHex(cid)
-    console.log(`multihash: ${multihash}`)
-
     // Add keyring to in-memory wallet
     const senderKeyring = caver.wallet.keyring.create(senderAddress, senderPrivateKey)
     caver.wallet.add(senderKeyring)
 
-    // Create ValueTransferMemo transaction
-    const tx = caver.transaction.valueTransferMemo.create({
+    // Create new private key
+    const newKey = caver.wallet.keyring.generateSingleKey()
+    console.log(`new private key: ${newKey}`)
+
+    // Create new Keyring instance with new private key
+    const newKeyring = caver.wallet.keyring.create(senderKeyring.address, newKey)
+    // Create an Account instance that includes the address and the public key
+    const account = newKeyring.toAccount()
+    console.log(account)
+
+    // Create account update transaction object
+    const accountUpdate = caver.transaction.accountUpdate.create({
+        from: senderKeyring.address,
+        account: account,
+        gas: 50000,
+    })
+
+    // Sign the transaction
+    await caver.wallet.sign(senderKeyring.address, accountUpdate)
+    // Send transaction
+    const receipt = await caver.rpc.klay.sendRawTransaction(accountUpdate)
+    console.log(`Account Update Transaction receipt => `)
+    console.log(receipt)
+
+    // Get accountKey from network
+    const accountKey = await caver.rpc.klay.getAccountKey(senderKeyring.address)
+    console.log(`Result of account key update to AccountKeyPublic`)
+    console.log(`Account address: ${senderKeyring.address}`)
+    console.log(`accountKey =>`)
+    console.log(accountKey)
+
+    // Update keyring with new private key in in-memory wallet
+    caver.wallet.updateKeyring(newKeyring)
+    // Send 1 Peb to recipient to test whether updated accountKey is well-working or not.
+    const vt = caver.transaction.valueTransfer.create({
         from: senderKeyring.address,
         to: recipientAddress,
         value: 1,
-        input: multihash,
-        gas: 30000,
+        gas: 25000,
     })
 
-    // Sign to the transaction
-    await caver.wallet.sign(senderKeyring.address, tx)
-
-    // Send a signed transaction to Klaytn
-    const receipt = await caver.rpc.klay.sendRawTransaction(tx)
-    console.log(receipt)
+    // Sign the transaction with updated keyring
+    await caver.wallet.sign(senderKeyring.address, vt)
+    // Send transaction
+    const vtReceipt = await caver.rpc.klay.sendRawTransaction(vt)
+    console.log(`Receipt of value transfer transaction after account update => `)
+    console.log(vtReceipt)
 }
