+++ ipfs/using_ipfs_with_caver/index.js
@@ -14,9 +14,9 @@ let senderPrivateKey = '' // e.g. '0x39a6375b608c2572fadb2ed9fd78c5c456ca3aa860c
 let recipientAddress = '' // e.g. '0xeb709d59954f4cdc6b6f3bfcd8d531887b7bd199'
 
 /**
- * Boilerplate code about "How to sign transaction using Wallet"
- * Related reference - Korean: https://ko.docs.klaytn.com/bapp/sdk/caver-js/getting-started#sending-a-value-transfer-transaction
- * Related reference - English: https://docs.klaytn.com/bapp/sdk/caver-js/getting-started#sending-a-value-transfer-transaction
+ * Example code about "Using IPFS with Caver."
+ * Related article - Korean: https://medium.com/klaytn/caver%EB%A5%BC-%EC%9D%B4%EC%9A%A9%ED%95%9C-ipfs-%EC%82%AC%EC%9A%A9%EB%B2%95-4889a3b29c0b
+ * Related article - English: https://medium.com/klaytn/using-ipfs-with-caver-964e1f721bfe
  */
 async function main() {
     try {
@@ -56,22 +56,44 @@ async function run() {
     }
     const caver = new Caver(new Caver.providers.HttpProvider(nodeApiUrl, option))
 
-    const senderKeyring = caver.wallet.keyring.create(senderAddress, senderPrivateKey)
+    // Set connection with IPFS Node
+    caver.ipfs.setIPFSNode('ipfs.infura.io', 5001, true)
+    // `ipfs.txt` is located at `caver-js-examples/ipfs/using_ipfs_with_caver/resources`.
+    const testFile = `${__dirname}/resources/ipfs.txt`
 
-    const vt = caver.transaction.valueTransfer.create({
+    // Add a file to IPFS with file path
+    const cid = await caver.ipfs.add(testFile)
+    console.log(`cid: ${cid}`)
+
+    // // Add a file to IPFS with file contents
+    // const contents = Buffer.from('test data')
+    // const cid = await caver.ipfs.add(contents)
+
+    // Download a file from IPFS
+    const buffer = await caver.ipfs.get(cid)
+    console.log(`Contents downloaded from IPFS: ${buffer.toString('utf8')}`)
+
+    // Convert from CID to multihash(hex formatted)
+    const multihash = await caver.ipfs.toHex(cid)
+    console.log(`multihash: ${multihash}`)
+
+    // Add keyring to in-memory wallet
+    const senderKeyring = caver.wallet.keyring.create(senderAddress, senderPrivateKey)
+    caver.wallet.add(senderKeyring)
+
+    // Create ValueTransferMemo transaction
+    const tx = caver.transaction.valueTransferMemo.create({
         from: senderKeyring.address,
         to: recipientAddress,
-        value: caver.utils.convertToPeb(1, caver.utils.klayUnit.KLAY.unit),
-        gas: 25000,
+        value: 1,
+        input: multihash,
+        gas: 30000,
     })
-    console.log('Before sign the transaction using wallet, there is no signatures.')
-    console.log(vt.signatures)
 
-    // Signing process
-    console.log('Add a senderKeyring to `caver.wallet`.')
-    caver.wallet.add(senderKeyring)
-    console.log('Sign the transaction using `caver.wallet`.')
-    await caver.wallet.sign(senderKeyring.address, vt)
-    console.log('Signature was added, so we can see the signature of sender.')
-    console.log(vt.signatures)
+    // Sign to the transaction
+    await caver.wallet.sign(senderKeyring.address, tx)
+
+    // Send a signed transaction to Klaytn
+    const receipt = await caver.rpc.klay.sendRawTransaction(tx)
+    console.log(receipt)
 }
