+++ setup.sh
@@ -7,7 +7,7 @@
 
 PROJECT_ROOT_DIR=$PWD
 ASSETS_DIR=$PROJECT_ROOT_DIR/assets
-BOILER_PLATE_TEMPLATE=$ASSETS_DIR/boilerplate.js
+BOILER_PLATE_TEMPLATE=$ASSETS_DIR/index.js
 NEW_SCENARIO=$1
 
 ## Tokenize $NEW_SCENARIO
