+++ transaction/send_klay_with_keystore_file/index.js
@@ -1,3 +1,4 @@
+const fs = require('fs')
 const path = require('path')
 const dotenv = require('dotenv')
 const Caver = require('caver-js')
@@ -11,11 +12,12 @@ let secretAccessKey = '' // e.g. 'aP/reVYHXqjw3EtQrMuJP4A3/hOb69TjnBT3ePKG'
 let chainId = '' // e.g. '1001' or '8217'
 let senderAddress = '' // e.g. '0xeb709d59954f4cdc6b6f3bfcd8d531887b7bd199'
 let senderPrivateKey = '' // e.g. '0x39a6375b608c2572fadb2ed9fd78c5c456ca3aa860c43192ad910c3269727fc1'
+let recipientAddress = '' // e.g. '0xeb709d59954f4cdc6b6f3bfcd8d531887b7bd199'
 
 /**
- * Boilerplate code about "How to create many EOAs easily."
- * Related reference - Korean: https://ko.docs.klaytn.com/klaytn/design/accounts#externally-owned-accounts-eoas
- * Related reference - English: https://docs.klaytn.com/klaytn/design/accounts#externally-owned-accounts-eoas
+ * Example code about "How to send KLAY with keystore file."
+ * Related article - Korean: https://medium.com/klaytn/common-architecture-of-caver-f7a7a1c554de
+ * Related article - English: https://medium.com/klaytn/common-architecture-of-caver-a714224a0047
  */
 async function main() {
     try {
@@ -40,6 +42,7 @@ function loadEnv() {
     chainId = chainId === '' ? envs.parsed.CHAIN_ID : chainId
     senderAddress = senderAddress === '' ? envs.parsed.SENDER_ADDRESS : senderAddress
     senderPrivateKey = senderPrivateKey === '' ? envs.parsed.SENDER_PRIVATE_KEY : senderPrivateKey
+    recipientAddress = recipientAddress === '' ? envs.parsed.RECIPIENT_ADDRESS : recipientAddress
 }
 
 async function run() {
@@ -54,11 +57,23 @@ async function run() {
     }
     const caver = new Caver(new Caver.providers.HttpProvider(nodeApiUrl, option))
 
-    const numEOAs = 10
-    const eoas = []
-    for (let i = 0; i < numEOAs; i++) {
-        eoas.push(caver.wallet.keyring.generate())
-    }
-    console.log(`Succeed to create ${numEOAs} EOAs using SingleKey.`)
-    console.log(eoas)
+    // 1. Create your own keystore file at "https://baobab.wallet.klaytn.com/create".
+    // 2. Rename that keystore file name with `keystore.json` or change the filename in line 63.
+    // 3. Place that keystore file at `caver-js-examples/kct/deploy_kip7_token_contract_with_keystore_file/resources`.
+    // 4. Get 5 KLAY at "https://baobab.wallet.klaytn.com/faucet".
+    const keystore = fs.readFileSync(`${__dirname}/resources/keystore.json`, 'utf8')
+    const password = '' // Put your password here.
+    const senderKeyring = caver.wallet.keyring.decrypt(keystore, password)
+    caver.wallet.add(senderKeyring)
+
+    // Send 1 KLAY.
+    const vt = caver.transaction.valueTransfer.create({
+        from: senderKeyring.address,
+        to: recipientAddress,
+        value: caver.utils.convertToPeb(1, caver.utils.klayUnit.KLAY.unit),
+        gas: 25000,
+    })
+    await caver.wallet.sign(senderKeyring.address, vt)
+    const receipt = await caver.rpc.klay.sendRawTransaction(vt)
+    console.log(receipt)
 }
