+++ wallet/sign_transaction/index.js
@@ -1,7 +1,6 @@
 const path = require('path')
 const dotenv = require('dotenv')
 const Caver = require('caver-js')
-const BigNumber = require('bignumber.js')
 
 const ROOT_DIR = path.join(__dirname, '../..') // Path can be changed based on its actual location.
 
@@ -15,9 +14,9 @@ let senderPrivateKey = '' // e.g. '0x39a6375b608c2572fadb2ed9fd78c5c456ca3aa860c
 let recipientAddress = '' // e.g. '0xeb709d59954f4cdc6b6f3bfcd8d531887b7bd199'
 
 /**
- * Boilerplate code about "How to use Klay Units."
- * Related reference - Korean: https://ko.docs.klaytn.com/klaytn/design/klaytn-native-coin-klay#units-of-klay
- * Related reference - English: https://docs.klaytn.com/klaytn/design/klaytn-native-coin-klay#units-of-klay
+ * Example code about "How to sign transaction using Wallet"
+ * Related reference - Korean: https://ko.docs.klaytn.com/bapp/sdk/caver-js/getting-started#sending-a-value-transfer-transaction
+ * Related reference - English: https://docs.klaytn.com/bapp/sdk/caver-js/getting-started#sending-a-value-transfer-transaction
  */
 async function main() {
     try {
@@ -58,35 +57,21 @@ async function run() {
     const caver = new Caver(new Caver.providers.HttpProvider(nodeApiUrl, option))
 
     const senderKeyring = caver.wallet.keyring.create(senderAddress, senderPrivateKey)
-    caver.wallet.add(senderKeyring)
 
-    // Because a field "value" always interprets its value as a unit "peb",
-    // you must take care what is the actual value when you sending some KLAY.
     const vt = caver.transaction.valueTransfer.create({
         from: senderKeyring.address,
         to: recipientAddress,
+        value: caver.utils.convertToPeb(1, caver.utils.klayUnit.KLAY.unit),
         gas: 25000,
-        value: 1,
     })
-    // Example-1: Sending 0.5 KLAY to recipient
-    // 1 KLAY is actually 10^18(=1000000000000000000) peb. So if you want send 0.5 KLAY,
-    // option-1: Set actual peb value directly to ValueTransfer transaction instance.
-    const bigNumberKLAY = new BigNumber('500000000000000000')
-    vt.value = bigNumberKLAY
-    console.log(`Example-1) The value what we set using option-1 is ${vt.value}`)
+    console.log('Before sign the transaction using wallet, there is no signatures.')
+    console.log(vt.signatures)
 
-    // option-2 (Recommended): option-1 is very cumbersome and difficult to read. Let's use more elegant way.
-    const KLAY = caver.utils.convertToPeb(5, caver.utils.klayUnit.KLAY.unit)
-    vt.value = KLAY / 10
-    console.log(`Example-1) The value what we set using option-2 is ${vt.value}`)
-
-    // Example-2: Sending 0.005 KLAY to recipient
-    // option-1: Use bigNumberKLAY we created at option-2 of Example-1.
-    vt.value = bigNumberKLAY / 100
-    console.log(`Example-2) The value what we set using option-1 is ${vt.value}`)
-
-    // option-2 (Recommended): option-1 has a high probability of being wrong by human error. Let's use more secure methomethodd.
-    const mKLAY = caver.utils.convertToPeb(5, caver.utils.klayUnit.mKLAY.unit)
-    vt.value = mKLAY
-    console.log(`Example-2) The value what we set using option-2 is ${vt.value}`)
+    // Signing process
+    console.log('Add a senderKeyring to `caver.wallet`.')
+    caver.wallet.add(senderKeyring)
+    console.log('Sign the transaction using `caver.wallet`.')
+    await caver.wallet.sign(senderKeyring.address, vt)
+    console.log('Signature was added, so we can see the signature of sender.')
+    console.log(vt.signatures)
 }
